scaffolding for parameters in service config
The idea here is to have a separate config file, params.pvn.yaml, which if present causes us to use the new service templating workflow. We will only register the template then have the user be able to push from UI and/or CI.

Check out `end_to_end_tests/configs/optional-services/with-parameters/` for example.

This PR just does the registration. Next step is to now wire this up to the UI.

 ayf

GitOrigin-RevId: c49e5381670335989d5a95507d1f8f60069d0e95