Track available replicas separately from created replicas
Done this way to increase chances of pod failure to trigger a change in
current state of the corresponding Deployment RuntimeObject. When a
change is detected, LogDebugInfo is called, which generates a
ProgramExit event for any failed pods.

This is not bullet proof. Its possible a pod fails and #replicas doesn't
change. To be super certain, we need to track list of pods and trigger
LogDebugInfo if pod names change.

 naphatkrit

GitOrigin-RevId: 883c4cb397d65fe503f71346c9497cf0d504f665